Is this the problem?
I am getting popup ads in this version. If this is what's causing it remove it please. I don't know for sure if this code is only for getting the data of add, and the code for displaying it is elsewhere. But, I would really like ads removed. Thank you.